# openapi.yaml
openapi: 3.0.0
info:
  title: Ridesharing API
  description: This is main API for our small ridesharing project
  contact:
    email: 1alekseik1@gmail.com
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: ref
servers:
- url: /
security:
- cookieAuth: []
...
paths:
...
  /get_multiple_users_info:
    post:
      tags:
        - user
      summary: Получить инфу о нескольких юзерах
      description: Все поля, _кроме пароля и __организации___. `organization_id` при этом возвращается.
        Если пользователь еще и водитель, то данные будут __только__ о пользователе, но не о водителе.
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                ids:
                  type: array
                  items:
                    type: integer
            examples:
              three ids:
                value:
                  ids: [1, 3, 5]
      responses:
        200:
          description: Все ок
          content:
            application/json:
              schema:
                type: array
                items:
                  allOf:
                    - $ref: '#/components/schemas/userID'
                    - $ref: '#/components/schemas/userBase'
                    - $ref: '#/components/schemas/userExtra'
        400:
          description: Любая ошибка в данных
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/errorBase'
              examples:
                main:
                  $ref: '#/components/examples/errorEmailNotFound'
        401:
          description: Нужна авторизация
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errorBase'
              examples:
                main:
                  $ref: '#/components/examples/errorAuthorizationRequred'
      security:
        - cookieAuth: []
  /upload_avatar:
    post:
      tags:
        - user
      summary: Поменять свой аватар
      description: Меняет только у текущего залогиненного пользователя
      requestBody:
        content:
          application/json:
            schema:
              required:
                - photo_url
              properties:
                photo_url:
                  type: string
      responses:
        200:
          description: Все ок, аватар сменился
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/userID'
                  - $ref: '#/components/schemas/userBase'
                  - $ref: '#/components/schemas/userExtra'
        400:
          description: Ошибка в запросе
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/errorBase'
        401:
          description: Нужна авторизация
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errorBase'
              examples:
                main:
                  $ref: '#/components/examples/errorAuthorizationRequred'
      security:
        - cookieAuth: []
...
  /change_email:
    post:
      tags:
        - user
      summary: Поменять свой email
      description: Меняет только у текущего залогиненного пользователя
      requestBody:
        content:
          application/json:
            schema:
              required:
                - email
              properties:
                email:
                  type: string
      responses:
        200:
          description: Все ок, email сменился
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/userID'
                  - $ref: '#/components/schemas/userBase'
                  - $ref: '#/components/schemas/userExtra'
        400:
          description: Ошибка в запросе
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/errorBase'
        401:
          description: Нужна авторизация
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errorBase'
              examples:
                main:
                  $ref: '#/components/examples/errorAuthorizationRequred'
      security:
        - cookieAuth: []
  /change_last_name:
    post:
      tags:
        - user
      summary: Поменять свою фамилию
      description: Меняет только у текущего залогиненного пользователя
      requestBody:
        content:
          application/json:
            schema:
              required:
                - last_name
              properties:
                last_name:
                  type: string
      responses:
        200:
          description: Все ок, фамилия сменилась
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/userID'
                  - $ref: '#/components/schemas/userBase'
                  - $ref: '#/components/schemas/userExtra'
        400:
          description: Ошибка в запросе
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/errorBase'
        401:
          description: Нужна авторизация
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errorBase'
              examples:
                main:
                  $ref: '#/components/examples/errorAuthorizationRequred'
      security:
        - cookieAuth: []
  /change_first_name:
    post:
      tags:
        - user
      summary: Поменять свое имя
      description: Меняет только у текущего залогиненного пользователя
      requestBody:
        content:
          application/json:
            schema:
              required:
                - first_name
              properties:
                first_name:
                  type: string
      responses:
        200:
          description: Все ок, имя сменилось
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/userID'
                  - $ref: '#/components/schemas/userBase'
                  - $ref: '#/components/schemas/userExtra'
        400:
          description: Ошибка в запросе
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/errorBase'
        401:
          description: Нужна авторизация
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errorBase'
              examples:
                main:
                  $ref: '#/components/examples/errorAuthorizationRequred'
      security:
        - cookieAuth: []
  /change_phone_number:
    post:
      tags:
        - user
      summary: Поменять свой номер телефона
      description: Меняет только у текущего залогиненного пользователя
      requestBody:
        content:
          application/json:
            schema:
              allOf:
                - $ref: '#/components/schemas/phoneNumber'
      responses:
        200:
          description: Все ок, номер сменился
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/userID'
                  - $ref: '#/components/schemas/userBase'
                  - $ref: '#/components/schemas/userExtra'
        400:
          description: Телефон не валиден
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/errorBase'
        401:
          description: Нужна авторизация
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errorBase'
              examples:
                main:
                  $ref: '#/components/examples/errorAuthorizationRequred'
      security:
        - cookieAuth: []
  /get_user_info:
    get:
      tags:
      - user
      summary: Получить всю инфу о User
      description: Все поля, кроме пароля. Если пользователь еще и водитель, то данные
        будут __только__ о пользователе, но не о водителе.
      responses:
        200:
          description: Все ок
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/userID'
                - $ref: '#/components/schemas/userBase'
                - $ref: '#/components/schemas/userExtra'
                required:
                  - cars
                properties:
                  cars:
                    type: array
                    items:
                      allOf:
                        - $ref: '#/components/schemas/car'
        400:
          description: Емэйл не найден
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/errorBase'
              examples:
                main:
                  $ref: '#/components/examples/errorEmailNotFound'
        401:
          description: Нужна авторизация
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errorBase'
              examples:
                main:
                  $ref: '#/components/examples/errorAuthorizationRequred'
      security:
      - cookieAuth: []
...
components:
  schemas:
...
    errorBase:
      required:
      - name
      - value
      type: object
      properties:
        name:
          type: string
        value:
          type: string
...
    car:
      allOf:
        - $ref: '#/components/schemas/carWithoutOwner'
      required:
        - owner_id
      properties:
        owner_id:
          type: integer
    carWithoutOwner:
      required:
        - model
        - color
        - registry_number
        - id
      properties:
        model:
          type: string
        color:
          type: string
        registry_number:
          type: string
        id:
          type: number
...
    userID:
      required:
      - id
      type: object
      properties:
        id:
          type: integer
    phoneNumber:
      required:
        - phone_number
      properties:
        phone_number:
          type: string
...
    userBase:
      required:
      - email
      - first_name
      - last_name
      - phone_number
      type: object
      properties:
        first_name:
          type: string
        last_name:
          type: string
        email:
          type: string
        organizations:
          $ref: '#/components/schemas/organizationInfo'
        phone_number:
          type: string
    userExtra:
      type: object
      properties:
        photo_url:
          type: string
          description: URL в хранилище
        is_driver:
          type: boolean
          description: Является ли пользователь водителем
...
    organizationInfo:
      required:
      - latitude
      - longitude
      - name
      type: object
      properties:
        id:
          type: integer
          description: ID в БД
        name:
          type: string
          description: Название организации
        latitude:
          type: number
          description: Координата широты
        longitude:
          type: number
          description: Координата долготы
        users:
          type: array
          description: Массив ID пользователей
          items:
            type: number
        address:
          type: string
          description: Адрес организации
        description:
          type: string
          description: Описание организации
        photo_url:
          type: string
          description: Фотка организации
...
  examples:
...
    errorEmailNotFound:
      value:
        name: Email not found
        value: mm.mm@mail.ru
...
    errorAuthorizationRequred:
      value:
        name: Authorization required
        value: ""
...
  securitySchemes:
    cookieAuth:
      type: apiKey
      description: Session=12d12jds
      name: session
      in: cookie